+++ src/lib/discord/thinking.ts
